src/requirements.txt
python-dotenv==1.0.1
fastapi==0.115.0
requests==2.32.3
pdfplumber==0.11.4
uvicorn==0.31.0
